[CIR][NFC] Consildate CIRGenExprCXX.cpp files
Somewhere in the upstreaming process, we created a file CIRGenCXXExpr.cpp
that corresponded to the file CIRGenExprCXX.cpp in the incubator. Later
we created a CIRGenExprCXX.cpp file.

This change consolidates those files, keeping the name used in the
incubator.
