+++ clang/lib/CIR/CodeGen/CMakeLists.txt
@@ -16,7 +16,6 @@ add_clang_library(clangCIR
   CIRGenCleanup.cpp
   CIRGenCXX.cpp
   CIRGenCXXABI.cpp
-  CIRGenCXXExpr.cpp
   CIRGenBuiltin.cpp
   CIRGenDecl.cpp
   CIRGenDeclOpenACC.cpp
